+++ examples/codegen/hello.c
@@ -3,10 +3,12 @@
 #include <inttypes.h>
 #include <stdbool.h>
 #include <string.h>
+#include <stddef.h>
+#include <pthread.h>
 
 
 int main(void) {
-    int64_t age = 41;
-    printf("%" PRId64 "\n", (int64_t)age);
+    int32_t age = 41;
+    printf("%d\n", (int)age);
     return 0;
 }
